pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "make-blueprint-creator"
version = "1.0.0"
authors = [
    {name = "AI Assistant", email = "ai@cursor.com"},
]
description = "Programmatic Make.com automation scenario creator and manager"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business :: Scheduling",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
keywords = [
    "make.com",
    "automation", 
    "workflow",
    "blueprint",
    "scenario",
    "api",
    "integration",
    "no-code",
    "low-code"
]
dependencies = [
    "requests>=2.28.0",
    "python-dotenv>=0.19.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=1.0.0",
    "safety>=2.0.0",
    "bandit>=1.7.0",
]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "coverage>=7.0.0",
]

[project.scripts]
make-blueprint = "make_blueprint_creator.cli.main:main"
make-examples = "make_blueprint_creator.cli.examples:main"
make-team-info = "make_blueprint_creator.cli.team_info:main"

[project.urls]
Homepage = "https://github.com/briankwest/make-blueprint-creator"
"Bug Reports" = "https://github.com/briankwest/make-blueprint-creator/issues"
Source = "https://github.com/briankwest/make-blueprint-creator"
Documentation = "https://github.com/briankwest/make-blueprint-creator#readme"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["make_blueprint_creator", "make_blueprint_creator.core", "make_blueprint_creator.utils", "make_blueprint_creator.examples", "make_blueprint_creator.cli"]

[tool.setuptools.package-data]
"make_blueprint_creator" = ["blueprints/*.json", "*.md"]
"*" = ["env.example"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.coverage.run]
source = ["src/make_blueprint_creator"]
omit = [
    "test_*.py",
    "run_tests.py",
    "setup.py",
    "*/venv/*",
    "*/__pycache__/*"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = [
    ".",
]
python_files = [
    "test_*.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
] 